Admin - Add new rights to access admin pages & a new publishers group

+++ lizmap/modules/admin/controllers/logs.classic.php
@@ -13,7 +13,7 @@ class logsCtrl extends jController
 {
     // Configure access via jacl2 rights management
     public $pluginParams = array(
-        '*' => array('jacl2.right' => 'lizmap.admin.access'),
+        '*' => array('jacl2.right' => 'lizmap.admin.lizmap.log.view'),
     );
 
     /**
@@ -101,13 +101,16 @@ class logsCtrl extends jController
         $rep = $this->getResponse('redirect');
 
         // Get counter
-        $cnx = jDb::getConnection('lizlog');
-
-        try {
-            $cnx->exec('DELETE FROM log_counter;');
-            jMessage::add(jLocale::get('admin~admin.logs.empty.ok', array('log_counter')));
-        } catch (Exception $e) {
-            jLog::log('Error while emptying table log_counter ');
+        if (\jAcl2::check('lizmap.admin.lizmap.log.delete')) {
+            try {
+                $cnx = jDb::getConnection('lizlog');
+                $cnx->exec('DELETE FROM log_counter;');
+                jMessage::add(jLocale::get('admin~admin.logs.empty.ok', array('log_counter')));
+            } catch (Exception $e) {
+                jLog::log('Error while emptying table log_counter ');
+            }
+        } else {
+            jMessage::add(jLocale::get('admin~admin.logs.error.no.delete.right'));
         }
 
         $rep->action = 'admin~logs:index';
@@ -228,13 +231,16 @@ class logsCtrl extends jController
         $rep = $this->getResponse('redirect');
 
         // Get counter
-        $cnx = jDb::getConnection('lizlog');
-
-        try {
-            $cnx->exec('DELETE FROM log_detail;');
-            jMessage::add(jLocale::get('admin~admin.logs.empty.ok', array('log_detail')));
-        } catch (Exception $e) {
-            jLog::log('Error while emptying table log_detail ');
+        if (\jAcl2::check('lizmap.admin.lizmap.log.delete')) {
+            try {
+                $cnx = jDb::getConnection('lizlog');
+                $cnx->exec('DELETE FROM log_detail;');
+                jMessage::add(jLocale::get('admin~admin.logs.empty.ok', array('log_detail')));
+            } catch (Exception $e) {
+                jLog::log('Error while emptying table log_detail ');
+            }
+        } else {
+            jMessage::add(jLocale::get('admin~admin.logs.error.no.delete.right'));
         }
 
         $rep->action = 'admin~logs:index';
@@ -252,12 +258,16 @@ class logsCtrl extends jController
         $rep = $this->getResponse('redirect');
 
         // Erase the log file
-        try {
-            $logPath = jApp::varPath('log/errors.log');
-            jFile::write($logPath, '');
-            jMessage::add(jLocale::get('admin~admin.logs.error.file.erase.ok', array('log_detail')));
-        } catch (Exception $e) {
-            jLog::log('Error while emptying the error log file');
+        if (\jAcl2::check('lizmap.admin.lizmap.log.delete')) {
+            try {
+                $logPath = jApp::varPath('log/errors.log');
+                jFile::write($logPath, '');
+                jMessage::add(jLocale::get('admin~admin.logs.error.file.erase.ok', array('log_detail')));
+            } catch (Exception $e) {
+                jLog::log('Error while emptying the error log file');
+            }
+        } else {
+            jMessage::add(jLocale::get('admin~admin.logs.error.no.delete.right'));
         }
 
         $rep->action = 'admin~logs:index';
